agent_config_security/requirements.txt
# Python Agent Config Security Requirements
# Python 3.10+

# Cryptography
cryptography>=41.0.0          # AES-256-GCM, PBKDF2
pyjwt[crypto]>=2.8.0          # JWT signing

# Key Management
azure-keyvault-secrets>=4.7.0  # Azure Key Vault (optional)
boto3>=1.28.0                 # AWS Secrets Manager (optional)
hvac>=2.1.0                   # HashiCorp Vault (optional)

# Security
pysecurecryptography>=0.1.0   # mlock support (Linux)
pynacl>=1.5.0                 # libsodium bindings

# Audit & Logging
structlog>=23.1.0             # Structured logging
python-json-logger>=2.0.7     # JSON logging

# Validation
pydantic>=2.0.0               # Config validation
python-dotenv>=1.0.0          # .env file loading

# Monitoring
prometheus-client>=0.17.0     # Metrics export
